+++ bentoml/__init__.py
@@ -24,8 +24,7 @@ from bentoml.version import __version__
 from bentoml.service import BentoService, api_decorator as api, \
     env_decorator as env, artifacts_decorator as artifacts
 from bentoml.server import metrics
-from bentoml.loader import load
-from bentoml.archive import save
+from bentoml.archive import save, load
 
 __all__ = [
     '__version__', 'api', 'env', 'artifacts', 'BentoService', 'save', 'load', 'handlers', 'metrics'
